check for distributed

+++ one-shot-tuning.py
@@ -353,7 +353,7 @@ def main(
             print(f"max memory allocated: {max_memory_allocated:.3f} GB.")
             with accelerator.accumulate(unet):
                 # Convert videos to latent space
-                pixel_values = batch["pixel_values"].to(weight_dtype)
+                pixel_values = batch["pixel_values"].to(weight_dtype).half()
                 video_length = pixel_values.shape[1]
                 pixel_values = rearrange(pixel_values, "b f c h w -> (b f) c h w")
                 check_gpu()
